+++ tests/tasks/test_task_assignment_notify_before_shooting.py
@@ -1,9 +1,9 @@
-"""Test Task to notify assignments 24hs before shooting."""
+"""Test Task to notify assignments before shooting."""
 from briefy.leica import models
 from briefy.leica.config import BEFORE_SHOOTING_SECONDS
-from briefy.leica.tasks.assignment import _notify_24hs_shooting
+from briefy.leica.tasks.assignment import _notify_before_shooting
 from briefy.leica.tasks.assignment import BEFORE_SHOOTING_MSG
-from briefy.leica.tasks.assignment import notify_24hs_shooting
+from briefy.leica.tasks.assignment import notify_before_shooting
 from conftest import BaseTaskTest
 from datetime import timedelta
 
@@ -11,7 +11,7 @@ import json
 
 
 class TestNotifyAssignmentBeforeShooting(BaseTaskTest):
-    """Test notify Assignment 24hs before shooting."""
+    """Test notify Assignment before shooting."""
 
     dependencies = [
         (models.Professional, 'data/professionals.json'),
@@ -32,7 +32,7 @@ class TestNotifyAssignmentBeforeShooting(BaseTaskTest):
         """
         assignment = instance_obj
         assignment.scheduled_datetime = now_utc + timedelta(seconds=3600)
-        notify_24hs_shooting()
+        notify_before_shooting()
 
         messages = self.get_messages_from_queue()
         assert len(messages) == 0
@@ -41,7 +41,7 @@ class TestNotifyAssignmentBeforeShooting(BaseTaskTest):
             seconds=int(BEFORE_SHOOTING_SECONDS) - 3600
         )
         assignment.state = 'pending'
-        notify_24hs_shooting()
+        notify_before_shooting()
 
         messages = self.get_messages_from_queue()
         assert len(messages) == 0
@@ -50,7 +50,7 @@ class TestNotifyAssignmentBeforeShooting(BaseTaskTest):
         """Will be successful if conditions are met.
 
         Assignment State is scheduled
-        Assignment now >= scheduled_datetime >= 24hs before the shooting
+        Assignment now >= scheduled_datetime >= BEFORE_SHOOTING_SECONDS
         """
         assignment = instance_obj
         assignment_id = assignment.id
@@ -58,15 +58,27 @@ class TestNotifyAssignmentBeforeShooting(BaseTaskTest):
         # do not notify with wrong status or scheduled datetime
         assignment.scheduled_datetime = now_utc + timedelta(seconds=3600)
         assignment.state = 'pending'
-        status = _notify_24hs_shooting(assignment)
+        status = _notify_before_shooting(assignment)
         assert status is False
 
-        # now we make sure status and scheduled datetime is correct
+        # now status is correct but date scheduled_datetime in the past
         assignment.state = 'scheduled'
-        assignment.scheduled_datetime = now_utc - timedelta(
+        assignment.scheduled_datetime = now_utc - timedelta(seconds=3600)
+        status = _notify_before_shooting(assignment)
+        assert status is False
+
+        # now scheduled_datetime is more than BEFORE_SHOOTING_SECONDS in seconds
+        assignment.scheduled_datetime = now_utc + timedelta(
+            seconds=int(BEFORE_SHOOTING_SECONDS) + 3600
+        )
+        status = _notify_before_shooting(assignment)
+        assert status is False
+
+        # finally we test a correct scheduled_datetime
+        assignment.scheduled_datetime = now_utc + timedelta(
             seconds=int(BEFORE_SHOOTING_SECONDS) - 3600
         )
-        status = _notify_24hs_shooting(assignment)
+        status = _notify_before_shooting(assignment)
 
         assert status is True
         messages = self.get_messages_from_queue()
@@ -74,12 +86,12 @@ class TestNotifyAssignmentBeforeShooting(BaseTaskTest):
         for message in messages:
             body = json.loads(message.body)
             assert body['event_name'] in (
-                'leica.task.notify_24hs_before_shooting.success'
+                'leica.task.notify_before_shooting.success'
             )
 
         assignment = models.Assignment.get(assignment_id)
         assert assignment.comments[0].content == BEFORE_SHOOTING_MSG
 
         # assert next time it will not notify again
-        status = _notify_24hs_shooting(assignment)
+        status = _notify_before_shooting(assignment)
         assert status is False