+++ HISTORY.rst
@@ -41,12 +41,16 @@ History
 
 2.1.28 (Unreleased)
 -------------------
+
     * New task to notify professionals about Assignments with late submissions (rudaporto).
     * New task to notify professionals before shooting (rudaporto).
+    * Card #462: two new _custom_filter for assignment endpoint: late_first_submission and late_re_submission (rudaporto).
+    * Card #462: two new _custom_filter for order endpoint: late_first_submission and late_re_submission (rudaporto).
 
 
 2.1.27 (2017-07-12)
 -------------------
+
     * Fixed Assignments report to return the category from the order (rudaporto).
     * Making sure that Order.actual_order_price is corrected updated after Order creation (rudaporto).